rename all modules
