+++ Woodoku/src/Entity/score_agent.py
