+++ Woodoku/src/UI/command_line_ui.py
@@ -1,8 +1,8 @@
 from typing import List, Tuple
 
-from Entity.WoodokuBoard import WoodokuBoard
-from Entity.WoodokuShape import WoodokuShape
-from UI.UIInterface import UIInterface
+from Entity.woodoku_board import WoodokuBoard
+from Entity.woodoku_shape import WoodokuShape
+from UI.ui_interface import UIInterface
 
 
 class CommandLineUI(UIInterface):
